Split types and encoding modules under file cap
